Clean up example
