+++ example/ios/Demo/LoadingIndicator/LoadingViewController.swift
@@ -11,13 +11,24 @@ import KotlinNativeFramework
 
 class LoadingViewController: UITableViewController {
 
-    private lazy var hudPresenter = HudPresenter(builder: HUD.Builder(viewController: self))
+    private lazy var viewModel = HudViewModel(builder: HUD.Builder(viewController: self))
+    private var lifecycleManager: LifecycleManager!
+    
+    deinit {
+        lifecycleManager.unbind()
+    }
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        
+        lifecycleManager = KNArchitectureFramework().bind(viewModel: viewModel, to: self) { _ in }
+    }
     
     override func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
         tableView.deselectRow(at: indexPath, animated: true)
         switch (indexPath.row) {
-        case 0: hudPresenter.showSystem()
-        case 1: hudPresenter.showCustom()
+        case 0: viewModel.onShowSystemPressed()
+        case 1: viewModel.onShowCustomPressed()
         default: ()
         }
     }
